+++ docs/scripts/routes/layout.js
@@ -1,7 +1,7 @@
 const TopAppBar = () => import('@/views/components/top-app-bar');
 const Grid = () => import('@/views/components/grid');
 const Form = () => import('@/views/components/form');
-const TextDivider = () => import('@/views/components/text-divider');
+const Divider = () => import('@/views/components/divider');
 
 export default [
   {
@@ -22,8 +22,8 @@ export default [
     component: Form
   },
   {
-    path: 'text-divider',
-    name: 'misc.text-divider',
-    component: TextDivider
+    path: 'divider',
+    name: 'misc.divider',
+    component: Divider
   }
 ];