+++ docs/scripts/views/components/divider.vue
@@ -1,10 +1,15 @@
 <template>
-  <ui-page name="text-divider" demo-count="2">
+  <ui-page name="divider" demo-count="2">
     <template #hero>
       <div class="hero-demo">
-        <ui-text-divider :type="typeOption">With Text</ui-text-divider>
+        <ui-divider v-show="withText" :type="typeOption">With Text</ui-divider>
+        <ui-divider v-show="!withText" :type="typeOption"></ui-divider>
       </div>
       <div class="hero-options">
+        <ui-form-field>
+          <ui-checkbox id="divider-with-text" v-model="withText"></ui-checkbox>
+          <label for="divider-with-text">With text</label>
+        </ui-form-field>
         <ui-select v-model="typeOption" class="hero-option" :options="TypeOptions">Type</ui-select>
       </div>
     </template>
@@ -13,15 +18,18 @@
     <section class="demo-wrapper">
       <h6 :class="$tt('headline6')">1.1 Default Usage</h6>
       <div class="demo">
-        <ui-text-divider>Text</ui-text-divider>
+        <ui-divider></ui-divider>
+      </div>
+      <div class="demo">
+        <ui-divider>Text</ui-divider>
       </div>
       <ui-snippet :code="$store.demos[1]"></ui-snippet>
     </section>
 
     <section class="demo-wrapper">
-      <h6 :class="$tt('headline6')">1.2 Vertical Divider</h6>
+      <h6 :class="$tt('headline6')">1.2 Vertical divider with text</h6>
       <div class="demo">
-        <ui-text-divider type="|" class="vertical-divider-demo">
+        <ui-divider type="|" class="vertical-divider-demo">
           <template #left>
             <div class="block red"></div>
           </template>
@@ -29,7 +37,7 @@
           <template #right>
             <div class="block blue"></div>
           </template>
-        </ui-text-divider>
+        </ui-divider>
       </div>
       <ui-snippet :code="$store.demos[2]"></ui-snippet>
     </section>
@@ -56,6 +64,7 @@ export default {
     return {
       // hero
       TypeOptions,
+      withText: false,
       typeOption: 0
     };
   }