+++ src/scripts/components/bottom-navigation.js
@@ -1,4 +1,4 @@
 import bootstrap from '../config/bootstrap';
-import UiBottomNavigation from './mobile/bottom-navigation';
+import UiBottomNavigation from './navigation/bottom-navigation';
 
 export default bootstrap(UiBottomNavigation);