+++ src/scripts/components/divider.js
@@ -1,0 +1,4 @@
+import bootstrap from '../config/bootstrap';
+import UiDivider from './others/divider';
+
+export default bootstrap(UiDivider);