+++ src/scripts/components/sheets/bottom-sheet.vue
@@ -1,0 +1,15 @@
+<template>
+  <div class="mdc-bottom-sheet">
+    <div class="mdc-bottom-sheet__scrim"></div>
+    <div class="mdc-bottom-sheet__content">
+      <slot></slot>
+    </div>
+  </div>
+</template>
+
+<script>
+// TODO
+export default {
+  name: 'UiBottomSheet'
+};
+</script>