+++ src/scripts/components/text-divider.js
@@ -1,4 +1,0 @@
-import bootstrap from '../config/bootstrap';
-import UiTextDivider from './dividers/text-divider';
-
-export default bootstrap(UiTextDivider);